+++ include/solver.hpp
@@ -2,9 +2,9 @@
  * GHOST (General meta-Heuristic Optimization Solving Tool) is a C++ framework
  * designed to help developers to model and implement optimization problem
  * solving. It contains a meta-heuristic solver aiming to solve any kind of
- * combinatorial and optimization real-time problems represented by a CSP/COP/EFSP/EFOP. 
+ * combinatorial and optimization real-time problems represented by a CSP/COP/EF-CSP/EF-COP. 
  *
- * First developped to solve game-related optimization problems, GHOST can be used for
+ * First developed to solve game-related optimization problems, GHOST can be used for
  * any kind of applications where solving combinatorial and optimization problems. In
  * particular, it had been designed to be able to solve not-too-complex problem instances
  * within some milliseconds, making it very suitable for highly reactive or embedded systems.
@@ -495,7 +495,7 @@ namespace ghost
 
 			if( is_optimization )
 			{
-				if( _best_opt_cost < 0 )
+				if( _model.objective->is_maximization() )
 				{
 					_best_opt_cost = -_best_opt_cost;
 					_cost_before_postprocess = -_cost_before_postprocess;