+++ src/objective.cpp
@@ -2,9 +2,9 @@
  * GHOST (General meta-Heuristic Optimization Solving Tool) is a C++ framework
  * designed to help developers to model and implement optimization problem
  * solving. It contains a meta-heuristic solver aiming to solve any kind of
- * combinatorial and optimization real-time problems represented by a CSP/COP/EFSP/EFOP. 
+ * combinatorial and optimization real-time problems represented by a CSP/COP/EF-CSP/EF-COP. 
  *
- * First developped to solve game-related optimization problems, GHOST can be used for
+ * First developed to solve game-related optimization problems, GHOST can be used for
  * any kind of applications where solving combinatorial and optimization problems. In
  * particular, it had been designed to be able to solve not-too-complex problem instances
  * within some milliseconds, making it very suitable for highly reactive or embedded systems.
@@ -58,7 +58,7 @@ double Objective::cost() const
 		throw nanException( _variables );
 
 	if( _is_maximization )
-		value *= -1;
+		value = -value;
 
 	return value;
 }
@@ -84,7 +84,7 @@ int Objective::expert_heuristic_value( const std::vector<Variable*>& variables,
 		simulated_cost = required_cost( variables );
 
 		if( _is_maximization )
-			simulated_cost *= -1;
+			simulated_cost = -simulated_cost;
 
 		if( min_cost > simulated_cost )
 		{